Fix regex pattern for camel_to_underscore. Code extracted from PR 110: https://github.com/vbabiy/djangorestframework-camel-case/pull/110
