+++ drf_camel_case/render.py
@@ -1,7 +1,7 @@
 from rest_framework.renderers import BrowsableAPIRenderer
 
-from djangorestframework_camel_case.settings import api_settings
-from djangorestframework_camel_case.util import camelize
+from drf_camel_case.settings import api_settings
+from drf_camel_case.util import camelize
 
 
 class CamelCaseJSONRenderer(api_settings.RENDERER_CLASS):
