+++ drf_camel_case/settings.py
