+++ setup.py
@@ -16,25 +16,25 @@ if sys.argv[-1] == "publish":
 
 readme = open("README.rst").read()
 history = open("HISTORY.rst").read().replace(".. :changelog:", "")
-import djangorestframework_camel_case
+import drf_camel_case
 
 setup(
-    name="djangorestframework-camel-case",
-    version=djangorestframework_camel_case.__version__,
+    name="drf-camel-case",
+    version=drf_camel_case.__version__,
     description="Camel case JSON support for Django REST framework.",
     long_description=readme + "\n\n" + history,
     long_description_content_type="text/x-rst",
     author="Vitaly Babiy",
     author_email="vbabiy86@gmail.com",
-    url="https://github.com/vbabiy/djangorestframework-camel-case",
-    packages=["djangorestframework_camel_case"],
-    package_dir={"djangorestframework_camel_case": "djangorestframework_camel_case"},
+    url="https://github.com/Novable-Open-Source/drf-camel-case",
+    packages=["drf_camel_case"],
+    package_dir={"drf_camel_case": "drf_camel_case"},
     include_package_data=True,
     python_requires=">=3.5",
     install_requires=[],
     license="BSD",
     zip_safe=False,
-    keywords="djangorestframework_camel_case",
+    keywords="drf_camel_case",
     classifiers=[
         "Development Status :: 2 - Pre-Alpha",
         "Intended Audience :: Developers",
