+++ tox.ini
@@ -11,7 +11,7 @@ python =
 
 [testenv]
 setenv =
-    PYTHONPATH = {toxinidir}:{toxinidir}/djangorestframework-camel-case
+    PYTHONPATH = {toxinidir}:{toxinidir}/drf-camel-case
 commands = python setup.py test
 deps =
     -r{toxinidir}/requirements.txt
